oop/inner-class

+++ oop/oop/src/main/java/org/cstefana/inheritance/Main.java
@@ -1,6 +1,7 @@
 package org.cstefana.inheritance;
 
 class Calc {
+
     public int add(int a, int b) {
         return a + b;
     }
@@ -10,8 +11,8 @@ class Calc {
     }
 }
 
-//inheritance
 class AdvCalc extends Calc {
+
     public double divide(double a, double b) {
         return a / b;
     }
@@ -21,16 +22,15 @@ class AdvCalc extends Calc {
     }
 }
 
-//multiple inheritance not exist but you call something like this
 class SuperAdvCalc extends AdvCalc {
-    public double root(double a){
+
+    public double root(double a) {
         return Math.sqrt(a);
     }
 }
 
 public class Main {
     public static void main(String[] args) {
-
         Calc calc = new Calc();
         System.out.println(calc.add(1, 2));
         System.out.println(calc.sub(1, 2));
@@ -38,14 +38,14 @@ public class Main {
         AdvCalc advCalc = new AdvCalc();
         System.out.println(advCalc.add(1, 2));
         System.out.println(advCalc.sub(1, 2));
-        System.out.println(advCalc.multiply(1, 2));
-        System.out.println(advCalc.divide(2, 2));
+        System.out.println(advCalc.multiply((double)1.0F, (double)2.0F));
+        System.out.println(advCalc.divide((double)2.0F, (double)2.0F));
 
         SuperAdvCalc superAdvCalc = new SuperAdvCalc();
         System.out.println(superAdvCalc.add(1, 2));
         System.out.println(superAdvCalc.sub(1, 2));
-        System.out.println(superAdvCalc.multiply(1, 2));
-        System.out.println(superAdvCalc.divide(2, 2));
-        System.out.println(superAdvCalc.root(2));
+        System.out.println(superAdvCalc.multiply((double)1.0F, (double)2.0F));
+        System.out.println(superAdvCalc.divide((double)2.0F, (double)2.0F));
+        System.out.println(superAdvCalc.root((double)2.0F));
     }
 }
